pyproject.toml
[tool.poetry]
name = "malabotpy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.1"
python-dotenv = "^0.17.0"
google-cloud-texttospeech = "^2.3.0"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
